create constants files
