+++ core-web/libs/portlets/dot-content-drive/portlet/src/lib/shared/models.ts
@@ -20,14 +20,3 @@ export interface DotContentDriveInit {
     path: string;
     filters: Record<string, string>;
 }
-
-export const SYSTEM_HOST: Site = {
-    identifier: 'SYSTEM_HOST',
-    hostname: 'SYSTEM_HOST',
-    type: 'HOST',
-    archived: false,
-    googleMap: ''
-};
-
-// We want to exclude forms and Hosts, and only show contentlets that are not deleted
-export const BASE_QUERY = '+systemType:false -contentType:forms -contentType:Host +deleted:false';
